+++ troncli/constants.py
@@ -2,14 +2,26 @@
 
 JAVA_TRON_RELEASES_URL = 'https://github.com/tronprotocol/java-tron/releases/download/'
 JAVA_TRON_LASTEST_VERSION = '3.2.1'
+JAVA_TRON_EVENT_NODE_GIT_URL = 'https://github.com/tronprotocol/java-tron.git'
+JAVA_TRON_EVENT_NODE_BRANCH_NAME = 'event_parser'
+
+TRON_GRID_GIT_URL = 'https://github.com/tronprotocol/tron-grid.git'
+TRON_GRID_BRANCH_NAME = 'master'
 
 NODES_DIR = '/tron_nodes'
+
 FULL_NODE_DIR = '/fullnode'
 SOLIDITY_NODE_DIR = '/soliditynode'
+EVENT_NODE_DIR = '/eventnode'
+GRID_API_DIR = '/gridapi'
+
 FULL_NODE_JAR = '/full.jar'
 SOLIDITY_NODE_JAR = '/solidity.jar'
+EVENT_NODE_JAR = '/event.jar'
+
 FULL_CONFIG = '/full.conf'
 SOL_CONFIG = '/sol.conf'
+EVENT_CONFIG = '/event.conf'
 
 LOCAL_HOST = '127.0.0.1:'
 
@@ -18,5 +30,8 @@ PRIVATENET_P2P_VERSION = 1
 
 MAINNET_ALLOW_CREATION_OF_CONTRACTS = 0
 PRIVATENET_ALLOW_CREATION_OF_CONTRACTS = 1
+
+DB_DISK_ONLY_VERSION = 1
+DB_IN_MEMORY_SUPPORT_VERSION = 2
 
 RUNNING_NODE_LIST_FILE = 'running_nodes.json'